backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
djangorestframework_simplejwt==5.5.1
python-decouple==3.8
dj-database-url==2.1.0
celery==5.3.4
redis==5.0.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
sqlparse==0.5.3
tzdata==2025.2
django-filter==23.5

# CV Parsing dependencies
PyPDF2==3.0.1
pdfplumber==0.11.7
python-docx==1.1.2
